requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
boto3==1.23.0
botocore==1.26.0
cffi==1.15.0
cryptography==37.0.2
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.6.0
jmespath==1.0.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
virtualenv==20.14.1
